Display message with URL on sora start
This is the url of the dashboard for the project that the device is in.
In the future, we may print multiple urls if the device is in multiple
projects, but right now, a device can only be in one project.

The url is hard coded for now as the hostname that is known by the
client is currently just the one for the grpc server, not the web
server. This will be fixed in this PR.