SORA-412 : [DRAFT] Got the persist-queue library working for saving device client data on to disk.  Replaced FIFO queues with SQLiteAckQueues for keeping the unacknowledged data in case of server exceptions.
Also making the _state_stream_sender thread to loop forever even in case of server exceptions.  This enables to persist the data on to disk in a simpler way instead of starting a new thread which could keep on restarting the thread until the server exceptions are fixed.

It took a while to make it working because on server exceptions, the queue acknowledgements were not being saved inorder to disk.